create clock component
